+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `createCalendarIcon()` & `createClockIcon()` in `icons.js` module
+
 ## [v0.0.14] - 2023-05-17
 
 ### Fixed